+++ lib/modules/generals/opaque/domain.d.ts
@@ -1,0 +1,4 @@
+declare const __indentifier__: unique symbol;
+export declare class NewType<identifier> {
+  private readonly [__indentifier__]: identifier;
+}
