+++ lib/modules/objects/domain.d.ts
@@ -1,11 +1,1 @@
-/** @deprecated use offical type PropertyKey instead */
-export type KeyOfObject = PropertyKey;
-
-declare global {
-  type uObject = unknownObject;
-  type unknownObject = { [key in PropertyKey]: unknown };
-
-  type anyObject = { [key in PropertyKey]: any };
-
-  type emptyObject = { [key in PropertyKey]: never };
-}
+export declare const empty_object_key: unique symbol;
